# gen/api/openapi.yaml
openapi: 3.0.2
info:
  title: IRS API
  version: "0.2"
servers:
- url: https://api.server.test/api/v0.2
paths:
  /irs/items/{globalAssetId}:
    post:
      description: Registers and starts a AAS crawler job for given {globalAssetId}
      operationId: getBomLifecycleByGlobalAssetId
      parameters:
      - description: Readable ID of manufacturer including plant
        explode: false
        in: path
        name: globalAssetId
        required: true
        schema:
          example: urn:uuid:6c311d29-5753-46d4-b32c-19b918ea93b0
          maxLength: 16
          minLength: 1
          pattern: urn:uuid:^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
          type: string
        style: simple
      - description: Unique identifier of a single, unique (sub)component/part/batch,
          given by its globalAssetId/ digital twin id
        explode: true
        in: query
        name: bomLifecycle
        required: true
        schema:
          enum:
          - asBuilt
          type: string
        style: form
      - description: Aspect information to add to the returned tree
        example: SerialPartTypization
        explode: false
        in: query
        name: aspect
        required: false
        schema:
          items:
            default: SerialPartTypization
            enum:
            - SerialPartTypization
            - AssemblyPartRelationship
            - PartDimension
            - SupplyRelationData
            - PCFCoreData
            - PCFTechnicalData
            - MarketPlaceOffer
            - MaterialAspect
            - BatteryPass
            - ProducDescription.Vehicle
            - ProducDescription.Battery
            - ReturnRequest
            - CertificateOfDestruction
            - CertificateOfDismantler
            - Adress
            - Contact
            type: string
          type: array
        style: form
      - description: Max depth of the returned tree, if empty max depth is returned
        explode: true
        in: query
        name: depth
        required: false
        schema:
          default: 1
          format: int32
          maximum: 100
          minimum: 0
          type: integer
        style: form
      - description: Direction in which the tree shall be traversed
        explode: true
        in: query
        name: direction
        required: false
        schema:
          default: downward
          enum:
          - upward
          - downward
          format: enum
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                complete:
                  $ref: '#/components/examples/job-id-response'
          description: job id response for successful job registration
        "400":
          content:
            application/json:
              examples:
                complete:
                  $ref: '#/components/schemas/ErrorResponse'
          description: Bad request
      summary: Get a BOM for a part
      tags:
      - Item Relationship Service
      x-accepts: application/json
  /irs/jobs/{jobId}:
    get:
      operationId: getBOMForJobId
      parameters:
      - description: 'Id of the job in registry. '
        explode: false
        in: path
        name: jobId
        required: true
        schema:
          format: uuid
          maxLength: 16
          minLength: 1
          type: string
        style: simple
      - description: If true, the endpoint returns the uncompleted results of the
          bom tree.
        explode: true
        in: query
        name: returnUncompletedResultTree
        required: false
        schema:
          default: true
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                complete:
                  $ref: '#/components/examples/complete-job-result'
          description: livecycle tree representation with the starting point of the
            given jobId
        "201":
          content:
            application/json:
              examples:
                complete:
                  $ref: '#/components/examples/job-result-without-uncompleted-result-tree'
          description: job details for given jobId
        "206":
          content:
            application/json:
              examples:
                complete:
                  $ref: '#/components/examples/partial-job-result'
          description: uncompleted livecycle tree representation with the starting
            point of the given jobId
        "404":
          content:
            application/json:
              examples:
                complete:
                  $ref: '#/components/examples/canceled-job-result'
          description: processing of job was canceled
        "417":
          content:
            application/json:
              example:
                complete:
                  $ref: '#/components/examples/failed-job-result'
          description: Processing of job failed
      tags:
      - Item Relationship Service
      x-accepts: application/json
  /irs/jobs/{processingState}:
    get:
      operationId: getJobsByProcessingState
      parameters:
      - description: List of jobs (globalAssetIds) for a certain processing state
        explode: false
        in: path
        name: processingState
        required: true
        schema:
          default: running
          enum:
          - running
          - failed
          - complete
          - canceled
          format: enum
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                complete:
                  $ref: '#/components/examples/complete-job-list-processing-state'
          description: list of jobs with given processingState
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
      tags:
      - Item Relationship Service
      x-accepts: application/json
  /irs/jobs/{jobId}/cancel:
    put:
      operationId: cancelJobForJobId
      parameters:
      - description: Id of the job in registry.
        explode: false
        in: path
        name: jobId
        required: true
        schema:
          example: 6c311d29-5753-46d4-b32c-19b918ea93b0
          format: uuid
          maxLength: 16
          minLength: 1
          type: string
        style: simple
      responses:
        "200":
          description: Job with {jobId} was canceled
        "400":
          description: Bad request. JobId must be a string in UUID format.
        "401":
          description: Authorization information is missing or invalid.
        "404":
          description: A job with the specified jobId was not found.
        "5XX":
          description: Unexpected error.
      tags:
      - Item Relationship Service
      x-accepts: application/json
components:
  examples:
    job-id-response:
      value:
        jobId: 6c311d29-5753-46d4-b32c-19b918ea93b0
    complete-job-list-processing-state:
      value:
        jobs:
        - 6c311d29-5753-46d4-b32c-19b918ea93b0
        - 46cd8fb1-34c1-4426-9c16-84b913bcfd95
    failed-job-result:
      value:
        job:
          jobId: e5347c88-a921-11ec-b909-0242ac120002
          globalAssetId: 6c311d29-5753-46d4-b32c-19b918ea93b0
          jobStatus: Failed
          createdOn: 2022-02-03T14:48:54.709Z
          startedOn: 2022-02-03T14:48:54.709Z
          lastModifiedOn: 2022-02-03T14:48:54.709Z
          jobCompleted: ""
          requestURL: https://api.server.test/api/../
          owner: ""
          summary:
            asyncFetchedItems:
              queued: 0
              running: 0
              complete: 0
              failed: 1
              lost: 0
          queryParameter:
            bomLifecycle: asBuilt
            aspect: SerialPartTypization
            depth: 4
            direction: downward
          exception:
            exception: IrsTimeoutException
            errorDetail: Timeout while requesting Digital Registry.
            exceptionDate: 2022-02-03T14:48:54.709Z
    partial-job-result-return-uncompleted-result-tree:
      value:
        job:
          jobId: e5347c88-a921-11ec-b909-0242ac120002
          globalAssetId: 6c311d29-5753-46d4-b32c-19b918ea93b0
          jobStatus: InProgress
          createdOn: 2022-02-03T14:48:54.709Z
          startedOn: 2022-02-03T14:48:54.709Z
          lastModifiedOn: 2022-02-03T14:48:54.709Z
          jobCompleted: 2022-02-03T14:48:54.709Z
          requestURL: https://api.server.test/api/../
          owner: ""
          summary:
            asyncFetchedItems:
              queued: 0
              running: 0
              complete: 0
              failed: 0
              lost: 0
          queryParameter:
            bomLifecycle: asBuilt
            aspect: SerialPartTypization
            depth: 4
            direction: downward
          exception:
            exception: IrsTimeoutException
            errorDetail: Timeout while requesting Digital Registry.
            exceptionDate: 2022-02-03T14:48:54.709Z
    partial-job-result:
      value:
        job:
          jobId: e5347c88-a921-11ec-b909-0242ac120002
          globalAssetId: 6c311d29-5753-46d4-b32c-19b918ea93b0
          jobStatus: InProgress
          createdOn: 2022-02-03T14:48:54.709Z
          startedOn: 2022-02-03T14:48:54.709Z
          lastModifiedOn: 2022-02-03T14:48:54.709Z
          jobCompleted: 2022-02-03T14:48:54.709Z
          requestURL: https://api.server.test/api/../
          owner: ""
          summary:
            asyncFetchedItems:
              queued: 0
              running: 0
              complete: 0
              failed: 0
              lost: 0
          queryParameter:
            bomLifecycle: asBuilt
            aspect: SerialPartTypization
            depth: 4
            direction: downward
          exception:
            exception: IrsTimeoutException
            errorDetail: Timeout while requesting Digital Registry.
            exceptionDate: 2022-02-03T14:48:54.709Z
    job-result-without-uncompleted-result-tree:
      value:
        job:
          jobId: e5347c88-a921-11ec-b909-0242ac120002
          globalAssetId: 6c311d29-5753-46d4-b32c-19b918ea93b0
          jobStatus: InProgress
          createdOn: 2022-02-03T14:48:54.709Z
          startedOn: 2022-02-03T14:48:54.709Z
          lastModifiedOn: 2022-02-03T14:48:54.709Z
          jobCompleted: 2022-02-03T14:48:54.709Z
          requestURL: https://api.server.test/api/../
          owner: ""
          summary:
            asyncFetchedItems:
              queued: 0
              running: 0
              complete: 0
              failed: 0
              lost: 0
          queryParameter:
            bomLifecycle: asBuilt
            aspect: SerialPartTypization
            depth: 4
            direction: downward
          exception:
            exception: IrsTimeoutException
            errorDetail: Timeout while requesting Digital Registry.
            exceptionDate: 2022-02-03T14:48:54.709Z
        relationships:
        - catenaXId: d9bec1c6-e47c-4d18-ba41-0a5fe8b7f447
          childItem:
            quantity:
              quantityNumber: 1
              measurementUnit:
                datatypeURI: urn:bamm:io.openmanufacturing:meta-model:1.0.0#piece
                lexicalValue: piece
            lifecycleContext: asBuilt
            assembledOn: 2022-02-03T14:48:54.709Z
            lastModifiedOn: 2022-02-03T14:48:54.709Z
            childCatenaXId: a45a2246-f6e1-42da-b47d-5c3b58ed62e9
        shells:
        - description:
          - language: en
            text: The shell for a vehicle
          globalAssetId:
            value:
            - a45a2246-f6e1-42da-b47d-5c3b58ed62e9
          idShort: future concept x
          identification: 882fc530-b69b-4707-95f6-5dbc5e9baaa8
          specificAssetIds:
          - key: engineserialid
            value: "12309481209312"
          submodelDescriptors:
          - description:
            - language: en
              text: Provides base vehicle information
            idShort: vehicle base details
            identification: 4a738a24-b7d8-4989-9cd6-387772f40565
            semanticId:
              value:
              - urn:bamm:com.catenax.vehicle:0.1.1
            endpoints:
            - interface: HTTP
              protocolInformation:
                endpointAddress: https://catena-x.net/vehicle/basedetails/
                endpointProtocol: HTTPS
                endpointProtocolVersion: "1.0"
          - description:
            - language: en
              text: Provides base vehicle information
            idShort: vehicle part details
            identification: dae4d249-6d66-4818-b576-bf52f3b9ae90
            semanticId:
              value:
              - urn:bamm:com.catenax.vehicle:0.1.1#PartDetails
            endpoints:
            - interface: HTTP
              protocolInformation:
                endpointAddress: https://catena-x.net/vehicle/partdetails/
                endpointProtocol: HTTPS
                endpointProtocolVersion: "1.0"
    canceled-job-result:
      value:
        job:
          jobId: e5347c88-a921-11ec-b909-0242ac120002
          globalAssetId: 6c311d29-5753-46d4-b32c-19b918ea93b0
          jobStatus: Canceled
          createdOn: 2022-02-03T14:48:54.709Z
          startedOn: 2022-02-03T14:48:54.709Z
          lastModifiedOn: 2022-02-03T14:48:54.709Z
          jobCompleted: 2022-02-03T14:48:54.709Z
          requestURL: https://api.server.test/api/../
          owner: ""
          summary:
            asyncFetchedItems:
              queued: 0
              running: 0
              complete: 0
              failed: 0
              lost: 0
          queryParameter:
            bomLifecycle: asBuilt
            aspect: SerialPartTypization
            depth: 4
            direction: downward
          exception:
            exception: IrsTimeoutException
            errorDetail: Timeout while requesting Digital Registry.
            exceptionDate: 2022-02-03T14:48:54.709Z
    complete-job-result:
      value:
        job:
          jobId: e5347c88-a921-11ec-b909-0242ac120002
          globalAssetId: 6c311d29-5753-46d4-b32c-19b918ea93b0
          jobStatus: Completed
          createdOn: 2022-02-03T14:48:54.709Z
          startedOn: 2022-02-03T14:48:54.709Z
          lastModifiedOn: 2022-02-03T14:48:54.709Z
          jobCompleted: 2022-02-03T14:48:54.709Z
          requestURL: https://api.server.test/api/../
          owner: ""
          summary:
            asyncFetchedItems:
              queued: 0
              running: 0
              complete: 0
              failed: 0
              lost: 0
          queryParameter:
            bomLifecycle: asBuilt
            aspect: SerialPartTypization
            depth: 4
            direction: downward
          exception:
            exception: IrsTimeoutException
            errorDetail: Timeout while requesting Digital Registry.
            exceptionDate: 2022-02-03T14:48:54.709Z
        relationships:
        - catenaXId: d9bec1c6-e47c-4d18-ba41-0a5fe8b7f447
          childItem:
            quantity:
              quantityNumber: 1
              measurementUnit:
                datatypeURI: urn:bamm:io.openmanufacturing:meta-model:1.0.0#piece
                lexicalValue: piece
            lifecycleContext: asBuilt
            assembledOn: 2022-02-03T14:48:54.709Z
            lastModifiedOn: 2022-02-03T14:48:54.709Z
            childCatenaXId: a45a2246-f6e1-42da-b47d-5c3b58ed62e9
        shells:
        - description:
          - language: en
            text: The shell for a vehicle
          globalAssetId:
            value:
            - a45a2246-f6e1-42da-b47d-5c3b58ed62e9
          idShort: future concept x
          identification: 882fc530-b69b-4707-95f6-5dbc5e9baaa8
          specificAssetIds:
          - key: engineserialid
            value: "12309481209312"
          submodelDescriptors:
          - description:
            - language: en
              text: Provides base vehicle information
            idShort: vehicle base details
            identification: 4a738a24-b7d8-4989-9cd6-387772f40565
            semanticId:
              value:
              - urn:bamm:com.catenax.vehicle:0.1.1
            endpoints:
            - interface: HTTP
              protocolInformation:
                endpointAddress: https://catena-x.net/vehicle/basedetails/
                endpointProtocol: HTTPS
                endpointProtocolVersion: "1.0"
          - description:
            - language: en
              text: Provides base vehicle information
            idShort: vehicle part details
            identification: dae4d249-6d66-4818-b576-bf52f3b9ae90
            semanticId:
              value:
              - urn:bamm:com.catenax.vehicle:0.1.1#PartDetails
            endpoints:
            - interface: HTTP
              protocolInformation:
                endpointAddress: https://catena-x.net/vehicle/partdetails/
                endpointProtocol: HTTPS
                endpointProtocolVersion: "1.0"
  schemas:
    BoMLifecycle:
      description: result object of tree representation
      type: object
    ErrorResponse:
      description: Error response
      properties:
        errors:
          description: List of errors
          items:
            description: List of errors
            type: string
          type: array
        message:
          description: Error message
          type: string
        statusCode:
          description: Error code
          enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
          type: string
      type: object